chore_: move backend fixtures to /tests, move config to /utils
